+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - **GM-455** - Changed layout style in footer
+### Added
+
+- **GM-397** - Filter mobile
 
 ## [0.8.2] - 2022-07-21